+++ events.go
@@ -104,7 +104,7 @@ func (e *eventDebouncer) debounce(frame frame) {
 		e.events = append(e.events, frame)
 	} else {
 		e.logger.Warning("Event buffer full, dropping event frame.",
-			newLogFieldString("event_name", e.name), newLogFieldStringer("frame", frame))
+			NewLogFieldString("event_name", e.name), NewLogFieldStringer("frame", frame))
 	}
 
 	e.mu.Unlock()
@@ -113,11 +113,11 @@ func (e *eventDebouncer) debounce(frame frame) {
 func (s *Session) handleEvent(framer *framer) {
 	frame, err := framer.parseFrame()
 	if err != nil {
-		s.logger.Error("Unable to parse event frame.", newLogFieldError("err", err))
+		s.logger.Error("Unable to parse event frame.", NewLogFieldError("err", err))
 		return
 	}
 
-	s.logger.Debug("Handling event frame.", newLogFieldStringer("frame", frame))
+	s.logger.Debug("Handling event frame.", NewLogFieldStringer("frame", frame))
 
 	switch f := frame.(type) {
 	case *schemaChangeKeyspace, *schemaChangeFunction,
@@ -128,7 +128,7 @@ func (s *Session) handleEvent(framer *framer) {
 		s.nodeEvents.debounce(frame)
 	default:
 		s.logger.Error("Invalid event frame.",
-			newLogFieldString("frame_type", fmt.Sprintf("%T", f)), newLogFieldStringer("frame", f))
+			NewLogFieldString("frame_type", fmt.Sprintf("%T", f)), NewLogFieldStringer("frame", f))
 	}
 }
 
@@ -181,7 +181,7 @@ func (s *Session) handleNodeEvent(frames []frame) {
 		switch f := frame.(type) {
 		case *topologyChangeEventFrame:
 			s.logger.Info("Received topology change event.",
-				newLogFieldString("frame", strings.Join([]string{f.change, "->", f.host.String(), ":", strconv.Itoa(f.port)}, "")))
+				NewLogFieldString("frame", strings.Join([]string{f.change, "->", f.host.String(), ":", strconv.Itoa(f.port)}, "")))
 			topologyEventReceived = true
 		case *statusChangeEventFrame:
 			event, ok := sEvents[f.host.String()]
@@ -199,7 +199,7 @@ func (s *Session) handleNodeEvent(frames []frame) {
 
 	for _, f := range sEvents {
 		s.logger.Info("Dispatching status change event.",
-			newLogFieldString("frame", strings.Join([]string{f.change, "->", f.host.String(), ":", strconv.Itoa(f.port)}, "")))
+			NewLogFieldString("frame", strings.Join([]string{f.change, "->", f.host.String(), ":", strconv.Itoa(f.port)}, "")))
 
 		// ignore events we received if they were disabled
 		// see https://github.com/apache/cassandra-gocql-driver/issues/1591
@@ -218,7 +218,7 @@ func (s *Session) handleNodeEvent(frames []frame) {
 
 func (s *Session) handleNodeUp(eventIp net.IP, eventPort int) {
 	s.logger.Info("Node is UP.",
-		newLogFieldStringer("event_ip", eventIp), newLogFieldInt("event_port", eventPort))
+		NewLogFieldStringer("event_ip", eventIp), NewLogFieldInt("event_port", eventPort))
 
 	host, ok := s.ring.getHostByIP(eventIp.String())
 	if !ok {
@@ -244,7 +244,7 @@ func (s *Session) startPoolFill(host *HostInfo) {
 
 func (s *Session) handleNodeConnected(host *HostInfo) {
 	s.logger.Debug("Pool connected to node.",
-		newLogFieldIp("host_addr", host.ConnectAddress()), newLogFieldInt("port", host.Port()), newLogFieldString("host_id", host.HostID()))
+		NewLogFieldIP("host_addr", host.ConnectAddress()), NewLogFieldInt("port", host.Port()), NewLogFieldString("host_id", host.HostID()))
 
 	host.setState(NodeUp)
 
@@ -255,7 +255,7 @@ func (s *Session) handleNodeConnected(host *HostInfo) {
 
 func (s *Session) handleNodeDown(ip net.IP, port int) {
 	s.logger.Warning("Node is DOWN.",
-		newLogFieldIp("host_addr", ip), newLogFieldInt("port", port))
+		NewLogFieldIP("host_addr", ip), NewLogFieldInt("port", port))
 
 	host, ok := s.ring.getHostByIP(ip.String())
 	if ok {